Software Engineering for AI/Module 5/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
joblib
sqlalchemy
psycopg2-binary
alembic
python-dotenv
transformers
torch
pymysql
